Fixes for chrome size of Win32Userwindow.
Take chrome size into account when setting the minimum and maximum
size. Record the chrome size for each window and only update it when
the style changes. Don't take into account the menu size in the chrome
-- this matches the behavior on Linux. When getting the window size,
get the width and height of the client area, not the window area.