Fix window.open for GtkUserWindow.
Set up the WebView configuration and show it only after the
web-view-ready signal is fired. This effectively splits UserWindow
initialization up into two steps. Lots of other minor cleanup and
fixes in GtkUserWindow.

Split UserWindow::GetBounds into GetBounds GetBoundsImpl, so that we
can expose a general version of it. Kill the platform-specific version
of UserWindow::GetId -- it's useless.