Fix OS X textured window background gradient.
When a large portion of a textured window is covered by an opaque
subview, OS X will push the window gradient up to emulate the look
of a non-textured window. Since we are *really* transparent, make the
alpha value of the view less than 1.0 to trick OS X into knowing that
we are transparent.

Also move some logic setting up the NSWindow from the WebView delegate
to the NSWindow subclass.
[#221 state:closed]