[TIDESK-574] Fix HTTPServer crashing issue.
Before we used the HTTPServerRequestHandler for the HTTPRequest
object we exposed into Kroll. The issue is this object's lifetime is
restricted by Poco to the handleRequest callback. Poco deletes this right after
this method returns. Once the GC of the interpreter runs it will try cleaning up
this HTTPRequest native object, but it's already deleted thus a crash results.

Moved HTTPRequest into it's own class to avoid the limited lifetime of the handler object.
Note: we may see issues with the request/response object if they are accessed outside the
callback's scope. Poco will most likely delete this after the handleRequest callback as well.